Refactor Unix interop
As we've discussed offline, we want to change how we're structuring
our interop code, in particular to make it easier to share interop
definitions across assemblies.  This commit is an initial refactoring
of our existing Unix interop code to follow the new approach.
