+++ src/Common/src/Interop/Unix/Interop.Libraries.cs
@@ -1,7 +1,10 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-internal partial class Interop
+internal static partial class Interop
 {
-    internal const string LIBC = "libc";
+    private static class Libraries
+    {
+        internal const string Libc = "libc";
+    }
 }
