+++ src/Common/src/Interop/Unix/libc/Interop.getcwd.cs
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Runtime.InteropServices;
+
+using size_t  = System.IntPtr;
+
+internal static partial class Interop
+{
+    internal static partial class libc
+    {
+        [DllImport(Libraries.Libc, SetLastError = true)]
+        internal static extern unsafe byte* getcwd(byte* buf, size_t bufSize);
+    }
+}
