+++ src/Common/src/Interop/Unix/libc/Interop.utime.cs
@@ -1,0 +1,22 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using System.Runtime.InteropServices;
+
+using time_t = System.IntPtr;
+
+internal static partial class Interop
+{
+    internal static partial class libc
+    {
+        [DllImport(Libraries.Libc, SetLastError = true)]
+        internal static extern int utime(string path, ref utimbuf times);
+
+        internal struct utimbuf 
+        { 
+           internal time_t actime; 
+           internal time_t modtime; 
+        }
+    }
+}
