+++ src/System.IO.FileSystem.DriveInfo/src/System.IO.FileSystem.DriveInfo.csproj
@@ -17,14 +17,14 @@
   </PropertyGroup>
   <!-- References are resolved from packages.config -->
   <ItemGroup Condition=" '$(OS)' == 'Unix' ">
-    <Compile Include="Interop\Interop.Linux.cs" />
     <Compile Include="System\IO\DriveInfo.Linux.cs" />
-    <Compile Include="$(CommonPath)\Interop\Interop.Errors.Unix.cs">
-      <Link>Common\Interop\Interop.Errors.Unix.cs</Link>
-    </Compile>
-    <Compile Include="$(CommonPath)\Interop\Interop.strerror.Unix.cs">
-      <Link>Common\Interop\Interop.strerror.Unix.cs</Link>
-    </Compile>
+    <Compile Include="$(CommonPath)\Interop\Unix\Interop.Libraries.cs"><Link>Interop\Interop.Libraries.cs</Link></Compile>
+    <Compile Include="$(CommonPath)\Interop\Unix\Interop.Errors.cs"><Link>Interop\Interop.Errors.cs</Link></Compile>
+    <Compile Include="$(CommonPath)\Interop\Unix\libc\Interop.statvfs.cs"><Link>Interop\Interop.statvfs.cs</Link></Compile>
+    <Compile Include="$(CommonPath)\Interop\Unix\libc\Interop.strerror.cs"><Link>Interop\Interop.strerror.cs</Link></Compile>
+    <Compile Include="$(CommonPath)\Interop\Linux\libc\Interop.endmntent.cs"><Link>Interop\Interop.endmntent.cs</Link></Compile>
+    <Compile Include="$(CommonPath)\Interop\Linux\libc\Interop.getmntent.cs"><Link>Interop\Interop.getmntent.cs</Link></Compile>
+    <Compile Include="$(CommonPath)\Interop\Linux\libc\Interop.setmntent.cs"><Link>Interop\Interop.setmntent.cs</Link></Compile>
   </ItemGroup>
   <ItemGroup Condition=" '$(OS)' == 'Windows_NT' ">
     <Compile Include="Interop\Interop.Windows.cs" />
