+++ src/System.IO.FileSystem/src/System/IO/UnixFileStream.cs
@@ -82,11 +82,11 @@ namespace System.IO
             _useAsyncIO = (options & FileOptions.Asynchronous) != 0;
 
             // Translate the arguments into arguments for an open call
-            Interop.OpenFlags openFlags = PreOpenConfigurationFromOptions(mode, access, options); // FileShare currently ignored
-            Interop.Permissions openPermissions = Interop.Permissions.S_IRWXU; // creator has read/write/execute permissions; no permissions for anyone else
+            Interop.libc.OpenFlags openFlags = PreOpenConfigurationFromOptions(mode, access, options); // FileShare currently ignored
+            Interop.libc.Permissions openPermissions = Interop.libc.Permissions.S_IRWXU; // creator has read/write/execute permissions; no permissions for anyone else
 
             // Open the file and store the safe handle. Subsequent code in this method expects the safe handle to be initialized.
-            _fileHandle = SafeFileHandle.Open(path, (int)openFlags, (int)openPermissions);
+            _fileHandle = SafeFileHandle.Open(path, openFlags, (int)openPermissions);
             _fileHandle.IsAsync = _useAsyncIO;
 
             // Lock the file if requested via FileShare.  This is only advisory locking. FileShare.None implies an exclusive 
@@ -94,8 +94,8 @@ namespace System.IO
             // and not atomic with file opening, it's better than nothing.
             try
             {
-                Interop.LockOperations lockOperation = (share == FileShare.None) ? Interop.LockOperations.LOCK_EX : Interop.LockOperations.LOCK_SH;
-                SysCall<int, int>((fd, op, _) => Interop.flock(fd, op), (int)(lockOperation | Interop.LockOperations.LOCK_NB));
+                Interop.libc.LockOperations lockOperation = (share == FileShare.None) ? Interop.libc.LockOperations.LOCK_EX : Interop.libc.LockOperations.LOCK_SH;
+                SysCall<Interop.libc.LockOperations, int>((fd, op, _) => Interop.libc.flock(fd, op), lockOperation | Interop.libc.LockOperations.LOCK_NB);
             }
             catch
             {
@@ -105,13 +105,13 @@ namespace System.IO
 
             // Support additional options after the file has been opened.
             // These provide hints around how the file will be accessed.
-            Interop.Advice fadv =
-                _options == FileOptions.RandomAccess ? Interop.Advice.POSIX_FADV_RANDOM :
-                _options == FileOptions.SequentialScan ? Interop.Advice.POSIX_FADV_SEQUENTIAL :
+            Interop.libc.Advice fadv =
+                _options == FileOptions.RandomAccess ? Interop.libc.Advice.POSIX_FADV_RANDOM :
+                _options == FileOptions.SequentialScan ? Interop.libc.Advice.POSIX_FADV_SEQUENTIAL :
                 0;
             if (fadv != 0)
             {
-                SysCall<int, int>((fd, advice, _) => Interop.posix_fadvise(fd, IntPtr.Zero, IntPtr.Zero, advice), (int)fadv);
+                SysCall<Interop.libc.Advice, int>((fd, advice, _) => Interop.libc.posix_fadvise(fd, IntPtr.Zero, IntPtr.Zero, advice), fadv);
             }
 
             // Jump to the end of the file if opened as Append.
@@ -161,10 +161,10 @@ namespace System.IO
         /// <param name="access">The FileAccess provided to the stream's constructor</param>
         /// <param name="options">The FileOptions provided to the stream's constructor</param>
         /// <returns>The flags value to be passed to the open system call.</returns>
-        private static Interop.OpenFlags PreOpenConfigurationFromOptions(FileMode mode, FileAccess access, FileOptions options)
+        private static Interop.libc.OpenFlags PreOpenConfigurationFromOptions(FileMode mode, FileAccess access, FileOptions options)
         {
             // Translate FileMode.  Most of the values map cleanly to one or more options for open/open64.
-            Interop.OpenFlags flags = 0;
+            Interop.libc.OpenFlags flags = 0;
             switch (mode)
             {
                 default:
@@ -173,19 +173,19 @@ namespace System.IO
 
                 case FileMode.Append: // Append is the same as OpenOrCreate, except that we'll also separately jump to the end later
                 case FileMode.OpenOrCreate:
-                    flags |= Interop.OpenFlags.O_CREAT;
+                    flags |= Interop.libc.OpenFlags.O_CREAT;
                     break;
 
                 case FileMode.Create:
-                    flags |= (Interop.OpenFlags.O_CREAT | Interop.OpenFlags.O_TRUNC);
+                    flags |= (Interop.libc.OpenFlags.O_CREAT | Interop.libc.OpenFlags.O_TRUNC);
                     break;
 
                 case FileMode.CreateNew:
-                    flags |= (Interop.OpenFlags.O_CREAT | Interop.OpenFlags.O_EXCL);
+                    flags |= (Interop.libc.OpenFlags.O_CREAT | Interop.libc.OpenFlags.O_EXCL);
                     break;
 
                 case FileMode.Truncate:
-                    flags |= Interop.OpenFlags.O_TRUNC;
+                    flags |= Interop.libc.OpenFlags.O_TRUNC;
                     break;
             }
 
@@ -193,15 +193,15 @@ namespace System.IO
             switch (access)
             {
                 case FileAccess.Read:
-                    flags |= Interop.OpenFlags.O_RDONLY;
+                    flags |= Interop.libc.OpenFlags.O_RDONLY;
                     break;
 
                 case FileAccess.ReadWrite:
-                    flags |= Interop.OpenFlags.O_RDWR;
+                    flags |= Interop.libc.OpenFlags.O_RDWR;
                     break;
 
                 case FileAccess.Write:
-                    flags |= Interop.OpenFlags.O_WRONLY;
+                    flags |= Interop.libc.OpenFlags.O_WRONLY;
                     break;
             }
 
@@ -216,7 +216,7 @@ namespace System.IO
                     break;
 
                 case FileOptions.WriteThrough:
-                    flags |= Interop.OpenFlags.O_SYNC;
+                    flags |= Interop.libc.OpenFlags.O_SYNC;
                     break;
             }
 
@@ -250,7 +250,7 @@ namespace System.IO
                 if (!_canSeek.HasValue)
                 {
                     // Lazily-initialize whether we're able to seek, tested by seeking to our current location.
-                    _canSeek = SysCall<int, int>((fd, _, __) => Interop.lseek64(fd, 0, (int)Interop.SeekWhence.SEEK_CUR), throwOnError: false) >= 0;
+                    _canSeek = SysCall<int, int>((fd, _, __) => Interop.libc.lseek64(fd, 0, Interop.libc.SeekWhence.SEEK_CUR), throwOnError: false) >= 0;
                 }
                 return _canSeek.Value;
             }
@@ -279,8 +279,8 @@ namespace System.IO
                 // Get the length of the file as reported by the OS
                 long length = SysCall<int, int>((fd, _, __) =>
                 {
-                    Interop.structStat stat;
-                    int result = Interop.fstat(fd, out stat);
+                    Interop.libc.structStat stat;
+                    int result = Interop.libc.fstat(fd, out stat);
                     return result >= 0 ? stat.st_size : result;
                 });
 
@@ -396,7 +396,7 @@ namespace System.IO
                     {
                         // Since we still have the file open, this will end up deleting
                         // it (assuming we're the only link to it) once it's closed.
-                        Interop.unlink(_path); // ignore any error
+                        Interop.libc.unlink(_path); // ignore any error
                     }
                 }
             }
@@ -444,7 +444,7 @@ namespace System.IO
         /// <summary>Flushes the OS buffer.  This does not flush the internal read/write buffer.</summary>
         private void FlushOSBuffer()
         {
-            SysCall<int, int>((fd, _, __) => Interop.fsync(fd));
+            SysCall<int, int>((fd, _, __) => Interop.libc.fsync(fd));
         }
 
         /// <summary>
@@ -567,7 +567,7 @@ namespace System.IO
                 SeekCore(value, SeekOrigin.Begin);
             }
 
-            SysCall<long, int>((fd, length, _) => Interop.ftruncate64(fd, length), value);
+            SysCall<long, int>((fd, length, _) => Interop.libc.ftruncate64(fd, length), value);
 
             // Return file pointer to where it was before setting length
             if (origPos != value)
@@ -662,7 +662,7 @@ namespace System.IO
             {
                 bytesRead = (int)SysCall((fd, ptr, len) =>
                 {
-                    long result = (long)Interop.read(fd, (byte*)ptr, (IntPtr)len);
+                    long result = (long)Interop.libc.read(fd, (byte*)ptr, (IntPtr)len);
                     Contract.Assert(result <= len);
                     return result;
                 }, (IntPtr)(bufPtr + offset), count);
@@ -798,7 +798,7 @@ namespace System.IO
             {
                 bytesWritten = SysCall((fd, ptr, len) =>
                 {
-                    long result = (long)Interop.write(fd, (byte*)ptr, (IntPtr)len);
+                    long result = (long)Interop.libc.write(fd, (byte*)ptr, (IntPtr)len);
                     Contract.Assert(result <= len);
                     return result;
                 }, (IntPtr)(bufPtr + offset), count);
@@ -958,7 +958,7 @@ namespace System.IO
             Contract.Assert(!_fileHandle.IsClosed && CanSeek);
             Contract.Assert(origin >= SeekOrigin.Begin && origin <= SeekOrigin.End);
 
-            long pos = SysCall((fd, off, or) => Interop.lseek64(fd, off, or), offset, (int)origin); // SeekOrigin values are the same as Interop.SeekWhence values
+            long pos = SysCall((fd, off, or) => Interop.libc.lseek64(fd, off, or), offset, (Interop.libc.SeekWhence)(int)origin); // SeekOrigin values are the same as Interop.libc.SeekWhence values
             _filePosition = pos;
             return pos;
         }
@@ -1002,7 +1002,7 @@ namespace System.IO
                     if (result < 0)
                     {
                         int errno = Marshal.GetLastWin32Error();
-                        if (errno == (int)Interop.Errors.EINTR)
+                        if (errno == Interop.Errors.EINTR)
                         {
                             continue;
                         }
